+++ lib/date/diff.ts
@@ -1,4 +1,4 @@
-import {isDate} from './is';
+import {convertToDate} from './convertToDate';
 
 export type DiffKey = 'week'
     | 'weeks'
@@ -22,22 +22,24 @@ const WEEK_IN_MILLISECONDS      = DAY_IN_MILLISECONDS * 7;
 /**
  * Compute the diff between two dates in the provided key
  *
- * @param {Date} val_a - Date to diff against
- * @param {Date} val_b - Date to diff with
+ * @param {Date|string} val_a - Date to diff against
+ * @param {Date|string} val_b - Date to diff with
  * @param {DiffKey} key - (default='millisecond') Key to diff in
  */
 function diff (
-    val_a:Date,
-    val_b:Date,
+    val_a:Date|string,
+    val_b:Date|string,
     key:DiffKey = 'millisecond'
 ):number {
+    const n_val_a = convertToDate(val_a);
+    const n_val_b = convertToDate(val_b);
     if (
-        !isDate(val_a) ||
-        !isDate(val_b)
+        n_val_a === null ||
+        n_val_b === null
     ) throw new TypeError('Diff requires date objects for both values');
 
     /* Get difference in milliseconds */
-    const diff_in_ms = val_a.valueOf() - val_b.valueOf();
+    const diff_in_ms = n_val_a.valueOf() - n_val_b.valueOf();
 
     switch (key) {
         case 'week':